refactor: 상품 구매 페이지에 이벤트를 추가하는 부분을 ProductPurchaseEvent.js로 분리
